Make window scale setting work properly
Add logical "resolution" property to Window to distinguish from the
physical "size" property.

Modify MainMenu to render correctly when resolution changes.

Remove high-DPI support for now, since it isn't essentially. Going to
add it back for the 1.5x scale which is a nice middle ground between
the 1x and 2x scales on Retina laptop screens.

Also set the default scale value to 2x.